requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
tavily-python
slowapi
markdown
jinja2